Added a few more lines that can cause cashes inside the game. Made it so the settings tool menu will not close now when enabling/disabling items. Fixed up some spelling mistakes. Addded a check if the unpacker can not find the datafil.big. Added a check to see if the unpacked folder exist and to create it if it dose not.

+++ DR2OTR_Randomizer/Resources/UnpackingAndPacking.cs
@@ -13,20 +13,26 @@ namespace DR2OTR_Randomizer.Resources
         string gamePath;
         public string Unpack(string path)
         {
-            Debug.WriteLine(workingPath);
+            if (!Directory.Exists(workingPath))
+            { Directory.CreateDirectory(workingPath); }
             //Copys the datafile.big from the games dicrotory
             //to be modfied then sets workingPath the unpacked folder
             //and returns it to be used for the randomizer
+            if (!File.Exists($"{path}\\data\\datafile.big"))
+            {   
+                MessageBox.Show("Could not find datafile.big. Make sure you have selected the right folder", "Warning");
+                return "";
+            }
             File.Copy($"{path}\\data\\datafile.big", $"{workingPath}\\datafile.big", true);
             var process = Process.Start($"{Application.StartupPath}\\Resources\\Unpacker\\Gibbed.DeadRising2.Unpack.exe"
                 , $"\"{workingPath}\\datafile.big\"");
             process.WaitForExit();
             gamePath = path;
-            var result = MessageBox.Show("Would you like to back up the datafile?", "Backup", MessageBoxButtons.YesNo);
+            var result = MessageBox.Show("Would you like to backup the datafile?", "Backup", MessageBoxButtons.YesNo);
             if (result == DialogResult.Yes)
             {
                 File.Copy($"{workingPath}\\datafile.big", $"{Application.StartupPath}\\datafile(Backup).big", true);
-                MessageBox.Show("The back up is located in the unpacked folder inside the Resources folder");
+                MessageBox.Show("The backup is located in the unpacked folder inside the Resources folder");
             }
             workingPath = $"{workingPath}\\datafile_big";
             return workingPath;
@@ -40,10 +46,10 @@ namespace DR2OTR_Randomizer.Resources
                 MessageBox.Show("Could not find datafile", "Warning");
                 return;
             }
-            var result = MessageBox.Show("This will overwrite the datafile inside of the games folder", "Warning", MessageBoxButtons.YesNo);
+            var result = MessageBox.Show("This will overwrite the datafile inside of the games folder. Continue?", "Warning", MessageBoxButtons.YesNo);
             if (result == DialogResult.No) 
             {
-                MessageBox.Show("The datafile as been packed and is located in the unpacked folder inside the resources folder. " +
+                MessageBox.Show("The datafile as has been packed and is located in the unpacked folder inside the resources folder." +
                     "You will need to rename it to \"datafile.big\" before adding it to your games directory");
                 return;
             }
